# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      active_workflows:
        description: |-
          ActiveWorkflows is populated on 409 workflow_resolution_required
          responses (§8.8). Count of open workflows blocking the mutation.
        example: 3
        type: integer
      allowed_actions:
        description: AllowedActions lists the P-26 resolution actions the caller may
          invoke.
        example:
        - replace_delegate
        - stop_workflows
        items:
          type: string
        type: array
      code:
        description: Code is a legacy alias for Error kept for backwards compatibility.
        example: not_found
        type: string
      details:
        description: Details is populated with per-field errors for 422 validation
          failures.
        items:
          $ref: '#/definitions/http.ValidationError'
        type: array
      error:
        description: Error is the machine-readable error code (LLD §17 error taxonomy).
        example: not_found
        type: string
      message:
        description: Message is a human-readable explanation of the error.
        example: resource not found
        type: string
      record_version:
        description: |-
          RecordVersion is populated on 409 optimistic_lock_conflict responses
          so the client can re-read and retry (CONC-4).
        example: 5
        type: integer
      request_id:
        description: RequestID is the per-request UUID injected by platform-gincommon.
        example: 01H8XYZ0000000000000000000
        type: string
      status:
        description: Status is the HTTP status code echoed in the body.
        example: 404
        type: integer
      trace_id:
        description: TraceID is the OTel W3C trace ID for cross-service correlation.
        example: a3f1b2c8d4e5f60718293a4b5c6d7e8f
        type: string
      workflow_ids:
        description: WorkflowIDs is populated on 409 workflow_resolution_required
          responses.
        items:
          type: string
        type: array
    type: object
...
  http.InternalLocaleResponse:
    properties:
      default_locale:
        example: en-US
        type: string
      tenant_id:
        format: uuid
        type: string
    type: object
  http.InternalMemberDeleteResponse:
    properties:
      removed:
        example: true
        type: boolean
      user_id:
        format: uuid
        type: string
    type: object
  http.InternalMembershipPatchRequest:
    properties:
      record_version:
        example: 1
        type: integer
      status:
        enum:
        - active
        - suspended
        type: string
    type: object
  http.InternalMembershipPatchResponse:
    properties:
      record_version:
        example: 2
        type: integer
      status:
        enum:
        - active
        - suspended
        type: string
      user_id:
        format: uuid
        type: string
    type: object
  http.InternalProvisionRequest:
    properties:
      default_locale:
        description: DefaultLocale is an optional BCP-47 locale (e.g. "en-US").
        example: en-US
        type: string
      name:
        description: Name is the display name.
        example: Acme Corp
        type: string
      owner_user_id:
        description: OwnerUserID is the initial tenant_owner.
        format: uuid
        type: string
      plan:
        description: Plan is the entitlement tier.
        enum:
        - starter
        - pro
        - enterprise
        type: string
      slug:
        description: Slug is the tenant slug (immutable once set).
        example: acme
        type: string
      tenant_id:
        description: TenantID is the pre-allocated tenant UUID.
        format: uuid
        type: string
    type: object
  http.InternalRealmPatchRequest:
    properties:
      keycloak_shard:
        example: shard-1
        type: string
      realm_id:
        example: acme
        type: string
      realm_type:
        enum:
        - shared
        - dedicated
        type: string
    type: object
  http.InternalTenantRealmResponse:
    properties:
      realm_id:
        type: string
      realm_type:
        enum:
        - shared
        - dedicated
        type: string
      tenant_id:
        format: uuid
        type: string
    type: object
  http.InternalTenderACLCheckResponse:
    properties:
      access_level:
        enum:
        - view
        - edit
        - approve
        type: string
      has_access:
        example: true
        type: boolean
    type: object
...
  http.SeatUsageResponse:
    properties:
      active_users:
        example: 12
        type: integer
      grace_ends_at:
        format: date-time
        type: string
      licensed_seats:
        example: 25
        type: integer
      over_cap:
        example: false
        type: boolean
      overage_since:
        format: date-time
        type: string
      pending_invitations:
        example: 3
        type: integer
    type: object
...
  http.TenantResponse:
    properties:
      default_locale:
        type: string
      id:
        type: string
      licensed_seats:
        type: integer
      local_accounts_enabled:
        type: boolean
      mfa_freshness_seconds:
        type: integer
      name:
        type: string
      overage_since:
        type: string
      ownerless_since:
        type: string
      plan:
        $ref: '#/definitions/domain.TenantPlan'
      realm_id:
        type: string
      realm_sync_pending:
        type: boolean
      realm_type:
        $ref: '#/definitions/domain.RealmType'
      record_version:
        type: integer
      slug:
        type: string
      status:
        $ref: '#/definitions/domain.SubscriptionStatus'
      subscription_started_at:
        type: string
      trial_ends_at:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: sharmila.dayalan@bcbpsolutions.com
    name: BCBP Solutions
  description: |-
    Organizational layer of the IAM subsystem — owns tenants, departments, memberships, tenant-level roles, delegations, invitations, ACL overlays, and group→role mappings.

    **Tenant isolation.** Every resource is scoped by `x-tenant-id`; cross-tenant reads are blocked by row-level security (RLS-1..RLS-6). The hot path `GET /internal/users/:id/memberships` is the enrichment source for AuthZ (SLO 15 ms cache-hit / 30 ms cache-miss).

    **Route prefixes.**  `/api/v1/*` — authenticated tenant users (JWT via gateway).  `/api/v1/internal/*` — in-mesh services (mTLS, `iam-system` role).  `/api/v1/operator/*` — platform operators (`platform_operator` role).

    Every mutation echoes `record_version` for optimistic-locking round-trip (CONC-4).
  license:
    name: Proprietary
  title: IAM Org & Membership API
  version: "1.0"
paths:
...
  /internal/tenants:
    post:
      consumes:
      - application/json
      description: Activates the 5 default system departments, creates 3 dept-role
        labels, grants tenant_owner, and emits TenantCreated + TrialStarted in one
        transaction. Called by the TrialTenantProvisioned consumer (not by external
        clients).
      parameters:
      - description: Provisioning payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.InternalProvisionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: slug_conflict OR tenant_already_provisioned
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-1 — Provision tenant (trial signup)
      tags:
      - internal
  /internal/tenants/{id}:
    patch:
      consumes:
      - application/json
      description: Called by Realm Provisioner after Keycloak realm creation. Sets
        the tenant's realm identity in a single UPDATE.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Realm identity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.InternalRealmPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.InternalTenantRealmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-2 — RP sets realm_id + realm_type + keycloak_shard atomically
      tags:
      - internal
  /internal/tenants/{id}/locale:
    get:
      description: Returns the tenant's default_locale for LLM-facing consumers.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.InternalLocaleResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-9 — Tenant default_locale (LLM Service)
      tags:
      - internal
  /internal/tenants/{id}/members/{user_id}:
    delete:
      description: Removes the user from the tenant, cascading dept memberships and
        role grants. Sets tenants.ownerless_since when the last owner is deleted (TM-12).
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.InternalMemberDeleteResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-5 — Delete member (full cascade + TM-12 ownerless_since)
      tags:
      - internal
    patch:
      consumes:
      - application/json
      description: Called by Realm Provisioner to relay KC user state (enabled/disabled)
        into tenant_memberships.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      - description: Lifecycle payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.InternalMembershipPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.InternalMembershipPatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: record_version mismatch (CONC-4)
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-4 — Keycloak lifecycle status update
      tags:
      - internal
  /internal/tenants/{id}/seat-usage:
    get:
      description: Same projection as P-27 but service-to-service (Billing pre-checks
        before charging for a new seat).
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.SeatUsageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-11 — Seat-usage S2S (Billing pre-check)
      tags:
      - internal
  /internal/tenants/{id}/tenders/{tender_id}/acl/{user_id}:
    get:
      description: Returns `{has_access, access_level}` — access_level is populated
        only when has_access is true.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tender UUID
        format: uuid
        in: path
        name: tender_id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.InternalTenderACLCheckResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-12 — Tender ACL check (active grant, TAE-3)
      tags:
      - internal
  /internal/tenants/{id}/tenders/{tender_id}/assignee-override:
    post:
      consumes:
      - application/json
      description: 422 assignee_ineligible per §16 A62 (NOT 409). Validates the new
        assignee is an active member at the required level in the department, then
        emits the event; persists nothing (OVR-1).
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tender UUID
        format: uuid
        in: path
        name: tender_id
        required: true
        type: string
      - description: Override payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.AssigneeOverrideRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.AssigneeOverrideResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: assignee_ineligible
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: I-13 — Validate-and-emit TenderAssigneeOverridden (OVR-1 no persistence)
      tags:
      - internal
  /internal/users/{id}/memberships:
    get:
      description: SLO 15 ms cache-hit / 30 ms miss (§21). Derived `member` role injected
        (TR-7). effective_feature_flags = planDefaults(plan) ⊕ tenants.feature_flags.
        Cache TTL 300s ± 30s jitter.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tenant UUID
        format: uuid
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: no active membership
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: 'I-8 — HOT PATH: full membership projection (AuthZ Enrichment)'
      tags:
      - internal
...
securityDefinitions:
  TenantID:
    description: Tenant UUID injected by the API gateway. Establishes the RLS scope
      for the request (RLS-6).
    in: header
    name: x-tenant-id
    type: apiKey
  TenantRoles:
    description: Comma-separated tenant-role list injected by the API gateway. Operator
      routes require `platform_operator`; internal routes require `iam-system`.
    in: header
    name: x-tenant-roles
    type: apiKey
  UserID:
    description: Authenticated user UUID injected by the API gateway. Use `iam-system`
      for in-mesh service calls hitting `/api/v1/internal/*`.
    in: header
    name: x-user-id
    type: apiKey
swagger: "2.0"